Added initial TargetCode class declarations.
Pretty much just work in progress, of course - not sure if it compiles at all. I developed it in a separate source tree and included on its own. It's possible it'll require merging with the C++ tree to make it work.

So far just about a third of instructions is declared and has method implementations.

Heavy use of inheritance, every instruction has its base in Instruction and implements printInstruction() method to obtain the resulting assembly code. Higher abstraction can be added easily in this phase.

To be done:

 * Implement all instructions

 * Map ICode -> TCode

 * Write register allocation

 * Fix bugs

I guess this is a good start - have been thinking about the final form for a few days total so I _hope_ the basic class layout will stay.

Comments incoming over time (especially when I'm bored on the bus)