Not that it does anything particularly useful but it 1) compiles && 2) doesn't crash....
Obviously, I got fed up with the ugly C macro infested API provided by SDCC, so I'm currently trying to bend it to my image.

I'm not really successful at that, to be honest. Current state is SDCC internals bitching about receiving NULL operands and such. I hope I'm fixing that soon...

Other news: The iCodes that aren't commented in tcode.cc are the ones that will be used in the demo compiler that should do the following things at least (consider it a basic functionality):

 * Call functions

 * Assign to variables (allocate registers)

 * Return from functions

 * Add variables, at least byte-sized